refactor transacation components out of transfer
